+++ docs/fr_FR/Changelog.rst
@@ -22,6 +22,7 @@ ChangeLog
 
 - Ikea RODRET dimmer: Mise-à-jour du modèle par défaut pour ON/OFF (2684).
 - Mise-à-jour FW PiZiGate: Amélioration scripts & support préliminaire PIv2 (2638).
+- Dump Zigate: Support préliminaire.
 
 240216-BETA-1
 -------------